3lab/requirements.txt

fastapi~=0.110
uvicorn[standard]~=0.29
pydantic~=2.7
pydantic-settings~=2.2
python-dotenv~=1.0
pyjwt~=2.8
passlib[bcrypt]~=1.7
sqlalchemy[aiosqlite]~=2.0
alembic~=1.13
celery~=5.4
redislite
rich~=13.7
httpx~=0.27
